remove ExploreListGrid and move FeaturedListCarousel
