--- packages/1ui/src/components/ExploreListGrid/ExploreListGrid.mdx
@@ -1,49 +1,0 @@
-import { Canvas, Controls, Meta, Source, Title } from '@storybook/blocks'
-
-import * as Stories from './ExploreListGrid.stories'
-
-<Meta of={Stories} />
-
-<Title />
-
-#### A grid layout optimized for the explore view of lists with wider spacing and responsive columns
-
-<Source
-  dark
-  language="tsx"
-  code={`
-import { ExploreListGrid } from '@0xintuition/1ui'
-
-const MyExploreGrid = () => (
-
-{' '}
-
-<ExploreListGrid>
-  <ListCard
-    displayName="Best Crypto Portfolio Trackers"
-    imgSrc="https://example.com/image.jpg"
-    identitiesCount={45}
-    buttonWrapper={(button) => (
-      <Link to="/lists/trackers" prefetch="intent">
-        {button}
-      </Link>
-    )}
-  />
-  <ListCard
-    displayName="Top DeFi Platforms"
-    imgSrc="https://example.com/image.jpg"
-    identitiesCount={32}
-    buttonWrapper={(button) => (
-      <Link to="/lists/defi" prefetch="intent">
-        {button}
-      </Link>
-    )}
-  />
-</ExploreListGrid>
-) `} />
-
-<Canvas of={Stories.BasicUsage} sourceState="none" />
-
-## Properties
-
-<Controls />
