--- packages/1ui/src/components/ExploreListGrid/index.ts
@@ -1,1 +1,0 @@
-export * from './ExploreListGrid'
